+++ app/controllers/accessibility_statement_controller.rb
@@ -1,5 +1,4 @@
 class AccessibilityStatementController < ApplicationController
-  include LocaleSwitchable
   before_action :update_locale
   def index; end
 end
