+++ app/controllers/admin/roles/permissions_controller.rb
@@ -1,6 +1,6 @@
 module Admin
   module Roles
-    class PermissionsController < AdminBaseController
+    class PermissionsController < ApplicationController
       before_action :authenticate_admin_user!
       layout 'admin'.freeze
