+++ app/controllers/admin_users/sessions_controller.rb
@@ -1,6 +1,5 @@
 module AdminUsers
   class SessionsController < Devise::SessionsController
-    include JourneyTypeIdentifiable
     def new
       @show_sign_in = Rails.configuration.x.admin_portal.show_form
       super
