+++ app/controllers/privacy_policy_controller.rb
@@ -1,5 +1,4 @@
 class PrivacyPolicyController < ApplicationController
-  include LocaleSwitchable
   before_action :update_locale
   def index; end
 end
