+++ app/controllers/problem_controller.rb
@@ -1,5 +1,4 @@
 class ProblemController < ApplicationController
-  include LocaleSwitchable
   before_action :update_locale
   def index; end
 end
