+++ app/controllers/saml_sessions_controller.rb
@@ -19,6 +19,7 @@ class SamlSessionsController < Devise::SamlSessionsController
   end
 
   def after_sign_in_path_for(_provider)
+    session[:journey_type] = :providers
     providers_confirm_office_path
   end
